Update font uses, add tag variant, and refactor ScheduledShiftsTr
