+++ frontend/src/App.tsx
@@ -2,7 +2,6 @@ import "@fontsource/inter/400.css";
 import "@fontsource/inter/500.css";
 import "@fontsource/inter/700.css";
 import "@fontsource/raleway/400.css";
-import "@fontsource/raleway/500.css";
 import "@fontsource/raleway/600.css";
 import "@fontsource/raleway/700.css";
 import "bootstrap/dist/css/bootstrap.min.css";
